dom/push/test/test_subscription_change.html

<!DOCTYPE HTML>
<html>
<!--
Bug 1205109: Make `pushsubscriptionchange` extendable.

Any copyright is dedicated to the Public Domain.
http://creativecommons.org/licenses/publicdomain/

-->
<head>
  <title>Test for Bug 1205109</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
  <meta http-equiv="Content-type" content="text/html;charset=UTF-8">
</head>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1205109">Mozilla Bug 1205109</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>

<script class="testbody" type="text/javascript">

  var registration;

  function start() {
    return navigator.serviceWorker.register("worker.js" + "?" + (Math.random()), {scope: "."})
    .then(swr => { registration = swr; return swr; });
  }

  var controlledFrame;
  function createControlledIFrame(swr) {
    var p = new Promise(function(res, rej) {
      var iframe = document.createElement('iframe');
      iframe.id = "controlledFrame";
      iframe.src = "/tests/dom/push/test/frame.html";

      iframe.onload = () => res();
      controlledFrame = iframe;
      document.body.appendChild(iframe);
    });
    return p;
  }

  function subscribe() {
    return registration.pushManager.subscribe();
  }

  function unsubscribe(pushSubscription) {
    controlledFrame.parentNode.removeChild(controlledFrame);
    controlledFrame = null;
    return pushSubscription.unsubscribe();
  }

  function unregister() {
    return registration.unregister();
  }

  function resetSubscriptions() {
    return new Promise((resolve, reject) => {
      var requestID = SpecialPowers.Cc["@mozilla.org/uuid-generator;1"]
                                   .getService(SpecialPowers.Ci
                                               .nsIUUIDGenerator)
                                   .generateUUID().toString();
      var cpmm = SpecialPowers.Cc["@mozilla.org/childprocessmessagemanager;1"]
                              .getService(SpecialPowers.Ci
                                          .nsISyncMessageSender);
      var listener = {
        receiveMessage: function(message) {
          message = SpecialPowers.Cu.waiveXrays(message);
          var isSuccess = message.name == "PushService:Reset:OK";
          var isFailure = !isSuccess && message.name == "PushService:Reset:KO";
          if ((isSuccess || isFailure) &&
              message.data.requestID == requestID) {

            cpmm.removeMessageListener("PushService:Reset:OK", listener);
            cpmm.removeMessageListener("PushService:Reset:KO", listener);
            (isSuccess ? resolve : reject)(message.data);
          }
        },
      };
      cpmm.addMessageListener("PushService:Reset:OK", listener);
      cpmm.addMessageListener("PushService:Reset:KO", listener);

      cpmm.sendAsyncMessage("Push:Reset", {
        requestID: requestID,
        scope: ".",
      }, null, SpecialPowers.Services.scriptSecurityManager
                            .getSystemPrincipal());
    });
  }

  function changeSubscription(oldSubscription) {
    return Promise.all([
      controlledFrame.contentWindow.waitOnWorkerMessage("changed"),
      resetSubscriptions(),
    ]).then(() =>
      registration.pushManager.getSubscription()
    ).then(newSubscription => {
      ok(newSubscription.endpoint != oldSubscription.endpoint,
        "Expected new push endpoint");

      var oldKey = new Uint8Array(oldSubscription.getKey("p256dh"));
      var newKey = new Uint8Array(newSubscription.getKey("p256dh"));
      ok(!oldKey.every((byte, index) => byte == newKey[index]),
        "Expected new key share");

      return newSubscription;
    });
  }

  function runTest() {
    start()
    .then(createControlledIFrame)
    .then(subscribe)
    .then(changeSubscription)
    .then(unsubscribe)
    .then(unregister)
    .catch(function(e) {
      ok(false, "Some test failed with error " + e);
    }).then(SimpleTest.finish);
  }

  SpecialPowers.pushPrefEnv({"set": [
    ["dom.push.enabled", true],
    ["dom.serviceWorkers.exemptFromPerDomainMax", true],
    ["dom.serviceWorkers.enabled", true],
    ["dom.serviceWorkers.testing.enabled", true]
    ]}, runTest);
  SpecialPowers.addPermission('push', true, document);
  SimpleTest.waitForExplicitFinish();
</script>
</body>
</html>
